Fix labelling by handling before conditionals
Recording the labels is easier when done before the conditionals that
whether we want the text represented side-by-side and whether we want
the outline represented in blocks. Previously we would check those
conditions and subsequently add labels whereever we thought appropriate
but that resulted to a format where the labels where positioned exactly
where the text is but still requires more maintenance as we have the
mark labels in more places.